client: Minor code style improvements. Apply gofmt -s.
Move debugging code out into a separate file.

Replace mixed use of fmt and log with just log.
