+++ client/main.go
@@ -1,10 +1,9 @@
-// Command client is the Instant Share client. It runs in your operating system's tray bar for quick access.
+// Instant Share client desktop app. It runs in your operating system's tray bar for quick access.
 package main
 
 import (
 	"bytes"
 	"flag"
-	"fmt"
 	"image"
 	"image/png"
 	"io/ioutil"
@@ -20,7 +19,6 @@ import (
 	_ "golang.org/x/image/tiff"
 )
 
-// TODO: Load from config. Have ability to set config.
 var hostFlag = flag.String("host", "", "Target server host.")
 var debugFlag = flag.Bool("debug", false, "Adds menu items for debugging purposes.")
 
@@ -47,7 +45,7 @@ func fileThumbnail(extension string, bytes []byte) trayhost.Image {
 // instantShareEnabled returns true if we have valid content in clipboard that can be instantly shared.
 // It also updates values of clipboard and notificationThumbnail if so.
 func instantShareEnabled() bool {
-	fmt.Println("grab content, content-type of clipboard")
+	log.Println("grab content, content-type of clipboard")
 
 	cc, err := trayhost.GetClipboardContent()
 	if err != nil {
@@ -96,7 +94,7 @@ func instantShareEnabled() bool {
 }
 
 func instantShareHandler() {
-	fmt.Println("request URL")
+	log.Println("request URL")
 
 	resp, err := httpClient.Get(*hostFlag + "/api/getfilename?ext=" + clipboard.extension)
 	if err != nil {
@@ -112,7 +110,7 @@ func instantShareHandler() {
 		return
 	}
 
-	fmt.Println("display/put URL in clipboard")
+	log.Println("display/put URL in clipboard")
 
 	url := *hostFlag + "/" + string(filename)
 	trayhost.SetClipboardText(url)
@@ -127,7 +125,7 @@ func instantShareHandler() {
 		},
 	}.Display()
 
-	fmt.Println("upload image in background of size", len(clipboard.bytes))
+	log.Println("upload image in background of size", len(clipboard.bytes))
 
 	go func(b []byte) {
 		req, err := http.NewRequest("PUT", url, bytes.NewReader(b))
@@ -142,74 +140,44 @@ func instantShareHandler() {
 			return
 		}
 		_ = resp.Body.Close()
-		fmt.Println("done")
+		log.Println("done")
 	}(clipboard.bytes)
 }
 
+func init() { log.SetFlags(0) }
+
+func init() { runtime.LockOSThread() }
+
 func main() {
 	flag.Parse()
-	runtime.LockOSThread()
 
 	menuItems := []trayhost.MenuItem{
-		trayhost.MenuItem{
+		{
 			Title:   "Instant Share",
 			Enabled: instantShareEnabled,
 			Handler: instantShareHandler,
 		},
 		trayhost.SeparatorMenuItem(),
-		trayhost.MenuItem{
+		{
 			Title:   "Quit",
 			Handler: trayhost.Exit,
 		},
 	}
 	if *debugFlag {
-		menuItems = append(menuItems,
-			trayhost.SeparatorMenuItem(),
-			trayhost.MenuItem{
-				Title: "Debug: Get Clipboard Content",
-				Handler: func() {
-					cc, err := trayhost.GetClipboardContent()
-					fmt.Printf("GetClipboardContent() error: %v\n", err)
-					fmt.Printf("Text: %q\n", cc.Text)
-					fmt.Printf("Image: %v len(%v)\n", cc.Image.Kind, len(cc.Image.Bytes))
-					fmt.Printf("Files: len(%v) %v\n", len(cc.Files), cc.Files)
-				},
-			},
-			trayhost.MenuItem{
-				Title: "Debug: Set Clipboard Text",
-				Handler: func() {
-					trayhost.SetClipboardText("http://www.example.com/image.png")
-				},
-			},
-			trayhost.MenuItem{
-				Title: "Debug: Notification",
-				Handler: func() {
-					handler := func() {
-						u4.Open("http://www.example.com/image.png")
-					}
-					notification := trayhost.Notification{Title: "Success", Body: "http://www.example.com/image.png", Timeout: 3 * time.Second, Handler: handler}
-					//trayhost.Notification{Title: "Upload Failed", Body: "error description goes here"}.Display()
-					if cc, err := trayhost.GetClipboardContent(); err == nil && cc.Image.Kind != "" {
-						notification.Image = cc.Image
-					}
-					notification.Display()
-				},
-			},
-		)
+		menuItems = append(menuItems, debugMenuItems()...)
 	}
 
-	// TODO: Create a real icon and bake it into the binary.
-	// TODO: Optionally, if non-Retina pixel perfection is required, generate or load 1x image and supply that as a second representation.
-	iconData, err := ioutil.ReadFile("./icon@2x.png")
+	// NOTE: Optionally, if non-Retina pixel perfection is required, generate or load 1x image and supply that as a second representation.
+	iconData, err := ioutil.ReadFile("icon@2x.png")
 	if err != nil {
 		panic(err)
 	}
 
-	fmt.Println("Starting.")
+	log.Println("Starting.")
 
 	trayhost.Initialize("Instant Share", iconData, menuItems)
 
 	trayhost.EnterLoop()
 
-	fmt.Println("Exiting.")
+	log.Println("Exiting.")
 }
